% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_quotes.R
\name{add_quotes}
\alias{add_quotes}
\title{Add escaped quotes around a string}
\usage{
add_quotes(x)
}
\arguments{
\item{x}{a string}
}
\value{
the same string but enclosed with quotes
}
\description{
This may be useful e.g. in preparing a Linux command to be
invoked via system()
}
\examples{
add_quotes("x")=='\"x\"'
}
