% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_age_from_dob.R
\name{get_age_from_dob}
\alias{get_age_from_dob}
\title{Calculate age}
\usage{
get_age_from_dob(dob, age_day = today(), units = "years", floor = TRUE)
}
\arguments{
\item{dob}{date-of-birth, the day to start calculating age.}

\item{age_day}{the date on which age is to be calculated.}

\item{units}{unit to measure age in. Defaults to \code{"years"}. Passed to \code{\link[lubridate:duration]{lubridate::duration()}}.}

\item{floor}{boolean for whether or not to floor the result. Defaults to \code{TRUE}.}
}
\value{
Age in \code{units}. Will be an integer if \code{floor = TRUE}.
}
\description{
By default, calculates the typical "age in years", with a
\code{floor} applied so that you are, e.g., 5 years old from
5th birthday through the day before your 6th birthday. Set
\code{floor = FALSE} to return decimal ages, and change \code{units}
for units other than years.
}
\examples{
my_dob <- as.Date('1983-10-20')
get_age_from_dob(my_dob)
get_age_from_dob(my_dob, units = "minutes")
get_age_from_dob(my_dob, floor = FALSE)
}
