% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_HSCP_Loc_DZ11_lookup.R
\name{get_HSCP_Loc_DZ11_lookup}
\alias{get_HSCP_Loc_DZ11_lookup}
\title{Get HSCP-Locality-DZ11 lookup file}
\usage{
get_HSCP_Loc_DZ11_lookup(
 
    file_path = "/conf/linkage/output/lookups/Unicode/Geography/HSCP Locality/HSCP Localities_DZ11_Lookup_20200825.rds"
)
}
\arguments{
\item{file_path}{string path to the lookup file}
}
\value{
a tibble holding the data from the lookup file
}
\description{
Get HSCP-Locality-DZ11 lookup file
}
\examples{
get_HSCP_Loc_DZ11_lookup()
}
