% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nice_colnames.R
\name{get_nice_colnames}
\alias{get_nice_colnames}
\title{get_nice_colnames
Rename columns according to named vector}
\usage{
get_nice_colnames(name_dict, r_names)
}
\arguments{
\item{name_dict}{A named vector mapping existing column names to new names}

\item{r_names}{Existing column names}
}
\value{

}
\description{
get_nice_colnames
Rename columns according to named vector
}
\examples{
my_col_dict <- c("mpg"="Miles per gallon")
names(mtcars) <- get_nice_colnames(names(mtcars),my_col_dict)
}
