% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pop_data.R
\name{get_pop_data}
\alias{get_pop_data}
\title{Get DataZone population file from cl-out}
\usage{
get_pop_data(
 
    pop_file_path = "/conf/linkage/output/lookups/Unicode/Populations/Estimates/DataZone2011_pop_est_2011_2019.rds"
)
}
\arguments{
\item{pop_file_path}{RDS file path}
}
\value{
Population data in long format in tibble
}
\description{
Get DataZone population file from cl-out
}
\examples{
df_pop <-get_pop_data("/conf/linkage/output/lookups/Unicode/Populations/Estimates/DataZone2011_pop_est_2011_2019.rds")
}
