% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_value_from_key.R
\name{get_value_from_key}
\alias{get_value_from_key}
\title{TBC}
\usage{
get_value_from_key(key_value_vector, key)
}
\arguments{
\item{key_value_vector}{A named vector}

\item{key}{A name in the named vector}
}
\value{

}
\description{
TBC
}
\examples{
get_value_from_key(c("x"="This is x"),"x")
}
