% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_head_file.R
\name{make_head_file}
\alias{make_head_file}
\title{Create small file from big one on file system
This function invokes Linux commands to create
a smaller file from a bigger one. This may be of use, e.g.,
to create a .csv file small enough to inspect in Excel from
a very large one.}
\usage{
make_head_file(big_file_path, n_rows = 10000)
}
\arguments{
\item{big_file_path}{the full path to a large file (e.g. .csv file)}

\item{n_rows}{the number of rows to be kept in new small file}
}
\value{
NULL (Console will display full path of small file or Linux response if command fails)
}
\description{
Create small file from big one on file system
This function invokes Linux commands to create
a smaller file from a bigger one. This may be of use, e.g.,
to create a .csv file small enough to inspect in Excel from
a very large one.
}
\examples{
big_file_path("/conf/LIST_analytics/Glasgow City/Training/TC/big_file.csv")
}
