% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/win_to_lin.R
\name{win_to_lin}
\alias{win_to_lin}
\title{Title}
\usage{
win_to_lin(win_path = "")
}
\arguments{
\item{win_path}{a windows file path}
}
\value{
a Linux path on the PHS stats/conf box
}
\description{
Title
}
\examples{
win_to_lin("\\\\\\\\Isdsf00d03\\\\LIST_analytics\\\\Glasgow City")
}
